+++ messages/en-US/media.php
@@ -24,6 +24,7 @@ return [
     'COMMON_FILES' => 'Files',
     'COMMON_FOLDER' => 'Folder',
     'COMMON_FOLDERS' => 'Folders',
+    'COMMON_RELATIONS' => 'Relations',
     'COMMON_REMOVE_TITLE' => 'Are you sure you want to remove this asset?',
     'COMMON_TRANSFORMATIONS' => 'Transformations',
     'EMBED_URL_EMBED_URL' => 'Embed URL',