+++ resources/views/admin/file/relations.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 /**
- * @see FileController::actionTransformations()
+ * @see FileController::actionRelations()
  *
  * @var View $this
  * @var File $file
@@ -11,11 +11,10 @@ declare(strict_types=1);
 
 use Hirtz\Media\Models\File;
 use Hirtz\Media\Modules\Admin\Controllers\FileController;
-use Hirtz\Media\Modules\Admin\Widgets\Grids\TransformationGridView;
+use Hirtz\Media\Modules\Admin\Widgets\Grids\FileRelationGridContainer;
 use Hirtz\Media\Modules\Admin\Widgets\Navs\FileHeader;
 use Hirtz\Media\Modules\Admin\Widgets\Navs\FileSubmenu;
 use Hirtz\Skeleton\Web\View;
-use Hirtz\Skeleton\Widgets\Grids\GridContainer;
 
 echo FileHeader::make()
     ->model($file);
@@ -23,7 +22,5 @@ echo FileHeader::make()
 echo FileSubmenu::make()
     ->model($file);
 
-echo GridContainer::make()
-    ->title(Yii::t('media', 'Transformations'))
-    ->grid(TransformationGridView::make()
-        ->file($file));
+echo FileRelationGridContainer::make()
+    ->file($file);